+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 Changes:
 
   * Added box type attribute to box sizes for specifying whether the item is a storage box or plate
+  * Autocomplete pasted values in bulk tables
   * improved performance of bulk library save
   * Fixed "Request header is too large" error limitting number of items that can be bulk created/edited
   * Fixed URL not saving correctly when editing SOPs